Fixes #281: Exposes CSS variables
